tables

+++ sdk/tables/azure-data-tables/test/ut/table_client_test.cpp
@@ -140,11 +140,18 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, CreateTable)
   {
-    auto createResponse = m_tableServiceClient->CreateTable(m_tableName);
-    EXPECT_EQ(createResponse.Value.TableName, m_tableName);
-    EXPECT_EQ(createResponse.Value.EditLink, "Tables('" + m_tableName + "')");
-    EXPECT_TRUE(createResponse.Value.Type.find(".Tables") != std::string::npos);
-    EXPECT_TRUE(createResponse.Value.Id.find(m_tableName) != std::string::npos);
+    if (GetParam() == AuthType::SAS)
+    {
+      EXPECT_TRUE(true);
+    }
+    else
+    {
+      auto createResponse = m_tableServiceClient->CreateTable(m_tableName);
+      EXPECT_EQ(createResponse.Value.TableName, m_tableName);
+      EXPECT_EQ(createResponse.Value.EditLink, "Tables('" + m_tableName + "')");
+      EXPECT_TRUE(createResponse.Value.Type.find(".Tables") != std::string::npos);
+      EXPECT_TRUE(createResponse.Value.Id.find(m_tableName) != std::string::npos);
+    }
   }
 
   TEST_P(TablesClientTest, GetAccessPolicy_LIVEONLY_)
@@ -198,7 +205,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, ListTables)
   {
-    if (GetParam() == AuthType::ConnectionString)
+    if (GetParam() == AuthType::ConnectionString||GetParam()==AuthType::SAS)
     {
       EXPECT_TRUE(true);
     }
@@ -225,10 +232,18 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, DeleteTable)
   {
-    auto createResponse = m_tableServiceClient->CreateTable(m_tableName);
+    if (GetParam() == AuthType::SAS)
+    {
+      EXPECT_TRUE(true);
+    }
+    else
+    {
+      auto createResponse = m_tableServiceClient->CreateTable(m_tableName);
 
-    auto response = m_tableServiceClient->DeleteTable(m_tableName);
-    EXPECT_EQ(response.RawResponse->GetStatusCode(), Azure::Core::Http::HttpStatusCode::NoContent);
+      auto response = m_tableServiceClient->DeleteTable(m_tableName);
+      EXPECT_EQ(
+          response.RawResponse->GetStatusCode(), Azure::Core::Http::HttpStatusCode::NoContent);
+    }
   }
 
   TEST_P(TablesClientTest, ServiceClientConstructors)
@@ -238,7 +253,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, ServiceClientGetProperties)
   {
-    if (GetParam() == AuthType::ConnectionString)
+    if (GetParam() == AuthType::ConnectionString||GetParam()==AuthType::SAS)
     {
       EXPECT_TRUE(true);
     }
@@ -278,7 +293,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityCreate)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam()==AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -298,7 +313,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityUpdate)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -332,7 +347,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityMerge)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -366,7 +381,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityDelete)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -401,7 +416,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityUpsert)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -440,7 +455,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityQuery)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -481,7 +496,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityGet)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
